# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users/{userId}/applications":{"get":{"tags":["application-access-controller"],"operationId":"listUserApplicationAccess","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApplicationAccessResponse"}}}}}},"put":{"tags":["application-access-controller"],"operationId":"replaceUserApplicationAccess","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplaceApplicationAccessRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApplicationAccessResponse"}}}}}}},"/api/v1/users/{id}/status":{"put":{"tags":["user-admin-controller"],"operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/users/{id}/role":{"put":{"tags":["user-admin-controller"],"operationId":"changeRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/users/me/password":{"put":{"tags":["user-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/applications/{id}":{"get":{"tags":["application-controller"],"operationId":"getApplicationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApplicationResponse"}}}}}},"put":{"tags":["application-controller"],"operationId":"updateApplication","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApplicationResponse"}}}}}}},"/api/v1/applications/{id}/status":{"put":{"tags":["application-controller"],"operationId":"changeStatus_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApplicationResponse"}}}}}}},"/api/v1/users":{"get":{"tags":["user-admin-controller"],"operationId":"listUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserResponse"}}}}}},"post":{"tags":["user-admin-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/users/{userId}/applications/{applicationId}":{"post":{"tags":["application-access-controller"],"operationId":"grantUserApplicationAccess","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationAccessGrantRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApplicationAccessResponse"}}}}}},"delete":{"tags":["application-access-controller"],"operationId":"removeUserApplicationAccess","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["authentication-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["authentication-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["authentication-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/applications":{"get":{"tags":["application-controller"],"operationId":"listApplications","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApplicationResponse"}}}}}},"post":{"tags":["application-controller"],"operationId":"createApplication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApplicationResponse"}}}}}}},"/api/v1/applications/{applicationId}/api-keys":{"get":{"tags":["application-api-key-controller"],"operationId":"listApiKeys","parameters":[{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApiKeyResponse"}}}}}},"post":{"tags":["application-api-key-controller"],"operationId":"createApiKey","parameters":[{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApiKeyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApiKeyCreationResponse"}}}}}}},"/api/v1/applications/{applicationId}/api-keys/{apiKeyId}/rotate":{"post":{"tags":["application-api-key-controller"],"operationId":"rotateApiKey","parameters":[{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"apiKeyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApiKeyCreationResponse"}}}}}}},"/api/v1/applications/{applicationId}/api-keys/{apiKeyId}/revoke":{"post":{"tags":["application-api-key-controller"],"operationId":"revokeApiKey","parameters":[{"name":"applicationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"apiKeyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["user-admin-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/applications/me":{"get":{"tags":["application-controller"],"operationId":"listVisibleApplications","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApplicationResponse"}}}}}}}},"components":{"schemas":{"ApplicationAccessGrantRequest":{"type":"object","properties":{"applicationId":{"type":"string","format":"uuid"},"accessLevel":{"type":"string","minLength":1}},"required":["accessLevel"]},"ReplaceApplicationAccessRequest":{"type":"object","properties":{"grants":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationAccessGrantRequest"}}},"required":["grants"]},"ApiResponseListApplicationAccessResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationAccessResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApplicationAccessResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"applicationId":{"type":"string","format":"uuid"},"applicationName":{"type":"string"},"applicationDisplayName":{"type":"string"},"accessLevel":{"type":"string"},"grantedBy":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseUserResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserResponse"},"timestamp":{"type":"string","format":"date-time"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string"},"status":{"type":"string"},"lastLoginAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UpdateUserRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":320,"minLength":0},"displayName":{"type":"string","maxLength":150,"minLength":0}},"required":["displayName","email"]},"ChangePasswordRequest":{"type":"object","properties":{"oldPassword":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":100,"minLength":6}},"required":["newPassword","oldPassword"]},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{},"timestamp":{"type":"string","format":"date-time"}}},"ApplicationRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"displayName":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string","maxLength":2000,"minLength":0}},"required":["displayName","name"]},"ApiResponseApplicationResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ApplicationResponse"},"timestamp":{"type":"string","format":"date-time"}}},"ApplicationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"displayName":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApplicationStatusRequest":{"type":"object","properties":{"status":{"type":"string","minLength":1}},"required":["status"]},"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":320,"minLength":0},"password":{"type":"string","maxLength":100,"minLength":6},"displayName":{"type":"string","maxLength":150,"minLength":0},"role":{"type":"string","minLength":1}},"required":["displayName","email","password","role"]},"ApiResponseApplicationAccessResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ApplicationAccessResponse"},"timestamp":{"type":"string","format":"date-time"}}},"RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"ApiResponseLoginResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginResponse"},"timestamp":{"type":"string","format":"date-time"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserResponse"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"CreateApiKeyRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"expiresAt":{"type":"string","format":"date-time"}},"required":["name"]},"ApiKeyCreationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"applicationId":{"type":"string","format":"uuid"},"name":{"type":"string"},"keyPrefix":{"type":"string"},"rawApiKey":{"type":"string"},"status":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseApiKeyCreationResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ApiKeyCreationResponse"},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListUserResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListApplicationResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiKeyResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"applicationId":{"type":"string","format":"uuid"},"name":{"type":"string"},"keyPrefix":{"type":"string"},"status":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"lastUsedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"revokedAt":{"type":"string","format":"date-time"}}},"ApiResponseListApiKeyResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ApiKeyResponse"}},"timestamp":{"type":"string","format":"date-time"}}}}}}